# bismarck-game/backend/docs/swagger.yaml
basePath: /api
definitions:
...
  internal_api_handlers.RefuelAllRequest:
    properties:
      fuel_amount:
        type: integer
      game_id:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: Bismarck Game Team
  description: API для игры Bismarck Chase - пошаговой стратегической игры о морских
    сражениях
  title: Bismarck Game API
  version: "1.0"
paths:
...
  /phases/complete:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные для завершения фазы
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Завершение текущей фазы
      tags:
      - Phases
  /phases/info:
    get:
      consumes:
      - application/json
      parameters:
      - description: Название фазы
        in: query
        name: phase
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Получение информации о фазе
      tags:
      - Phases
  /phases/next:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные для перехода к следующей фазе
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Переход к следующей фазе
      tags:
      - Phases
  /phases/records:
    get:
      consumes:
      - application/json
      parameters:
      - description: ID игры
        in: query
        name: game_id
        required: true
        type: string
      - description: Номер хода
        in: query
        name: turn
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Получение записей фаз
      tags:
      - Phases
  /phases/start:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные для запуска фазы
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Запуск новой фазы
      tags:
      - Phases
  /phases/turn/start:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные для запуска хода
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Запуск нового хода
      tags:
      - Phases
  /refuel/all:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные для заправки
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api_handlers.RefuelAllRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Заправка всех кораблей в игре
      tags:
      - Refuel
...
securityDefinitions:
  Bearer:
    description: 'JWT токен в формате: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"